+++ ndk/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - ndk/hardware_buffer: Make `HardwareBuffer::as_ptr()` public for interop with Vulkan.
+- **Breaking:** `Configuration::country()` now returns `None` when the country is unset (akin to `Configuration::language()`)
 - Added `media/media_codec` bindings
 
 # 0.6.0 (2022-01-05)